Fixing some endpoints and adding notifications

+++ src/main/java/com/app/usochicamochabackend/review/infrastructure/entity/InspectionEntity.java
@@ -32,7 +32,7 @@ public class InspectionEntity {
     private Boolean unexpected;
     @Column(nullable = false)
     private LocalDateTime dateStamp;
-    private Integer hourMeter;
+    private Double hourMeter;
     private String leakStatus;
     private String brakeStatus;
     private String beltsPulleysStatus;
