+++ src/main/java/com/app/usochicamochabackend/update/application/dto/PerformChangeMotorOilRequest.java
@@ -8,6 +8,6 @@ public record PerformChangeMotorOilRequest(
         LocalDateTime dateTime,
         String brand,
         Integer quantity,
-        Integer currentHourMeter,
+        Double currentHourMeter,
         Integer averageHoursChange
 ) {}
