+++ src/main/java/com/app/usochicamochabackend/update/application/dto/PerformChangeMotorOilResponse.java
@@ -11,6 +11,6 @@ public record PerformChangeMotorOilResponse(
         LocalDateTime timestamp,
         String brand,
         Integer quantity,
-        Integer currentHourMeter,
+        Double currentHourMeter,
         Integer averageHoursChange
 ) {}
